Nested LiveViews - Forms targets behave differently in tests
If you run this test if will fail
`mix test test/liveview_nested_form_test_web/live/page_live_test.exs`
because the parent LiveView does not have a
handle_event for "SUBMIT".

I have tried adding a phx-target manually but that
doesn't seem to help either.

It works as expected in the browser. The "SUBMIT"
event is sent to the nested LiveView. But in the
test the event is sent to the parent.